% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFSreconstruction.R
\name{AFSreconstruction}
\alias{AFSreconstruction}
\title{Advancing front surface reconstruction}
\usage{
AFSreconstruction(points, jetSmoothing = NULL)
}
\arguments{
\item{points}{numeric matrix which stores the points, one point per row}

\item{jetSmoothing}{if not \code{NULL}, must be an integer, and then the 
points cloud is smoothed before the reconstruction, using this integer as 
the number of neighbors for the smoothing; note that this smoothing 
relocates the points and then should not be used if the points have been 
sampled without noise on the surface}
}
\value{
A \code{cgalMesh} object.
}
\description{
Reconstruction of a surface from a cloud of 3D points.
}
\details{
See \href{https://doc.cgal.org/latest/Advancing_front_surface_reconstruction/index.html#Chapter_Advancing_Front_Surface_Reconstruction}{Advancing Front Surface Reconstruction}.
}
\examples{
library(cgalMeshes)
data(bunny, package = "onion")
mesh <- AFSreconstruction(bunny)
rglMesh <- mesh$getMesh()
library(rgl)
shade3d(rglMesh, color = "firebrick")
}
