% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isosurface.R
\name{algebraicMeshesIntersection}
\alias{algebraicMeshesIntersection}
\title{Intersection of algebraic surfaces}
\usage{
algebraicMeshesIntersection(
  polynomials,
  sphereCenter,
  sphereRadius,
  angleBound,
  radiusBound,
  distanceBound,
  errorBound = 0.001
)
}
\arguments{
\item{polynomials}{the polynomials defining the isosurfaces, a list; 
each element of this list can be either a \code{\link[spray]{spray}} 
object or a list with two fields: \code{exponents}, an integer matrix with 
three columns, and \code{coeffs}, a numeric vector; each isosurface is 
then defined by \code{P(x,y,z)=0}, where \code{P} is a polynomial of this 
list}

\item{sphereCenter, sphereRadius}{center and radius of a sphere bounding the 
isosurface; the values of the polynomials at the center of this sphere 
must be less than zero}

\item{angleBound}{lower bound in degrees for the angles of the faces of 
the mesh}

\item{radiusBound}{upper bound for the radii of surface Delaunay balls; 
 a surface Delaunay ball is a ball circumscribing a face and centered 
at the surface}

\item{distanceBound}{upper bound for the distance between the 
circumcenter of a face and the center of the surface Delaunay 
ball of this face}

\item{errorBound}{a relative error bound used in the computations}
}
\value{
A \code{cgalMesh} object.
}
\description{
Computes a mesh of the intersection of some algebraic surfaces 
  (isosurfaces defined by a polynomial).
}
\note{
If either \code{angleBound} is too high, \code{radiusBound} is 
  too small, or \code{distanceBound} is too small, the computation could 
  never finishes. Therefore, it is recommended to start with a small 
  \code{angleBound} and large \code{radiusBound} and \code{distanceBound}, 
  and then to adjust these parameters if the resulting mesh is not nice.
}
