% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someMeshes.R
\name{conformalCyclideMesh}
\alias{conformalCyclideMesh}
\title{Conformal cyclide mesh}
\usage{
conformalCyclideMesh(a, c, aspectRatio, normals = TRUE, nu = 90L, nv = 40L)
}
\arguments{
\item{a, c}{cyclide parameters, \code{a > c > 0}; see the figure in the
documentation of \code{\link{cyclideMesh}}}

\item{aspectRatio}{the aspect ratio, positive number}

\item{normals}{Boolean, whether to add normals to the mesh}

\item{nu, nv}{numbers of subdivisions, integers (at least 3)}
}
\value{
A triangle \strong{rgl} mesh (class \code{mesh3d}).
}
\description{
Cyclide mesh of a Dupin cyclide constructed from a 
  conformal parameterization.
}
