% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexHull.R
\name{convexHull}
\alias{convexHull}
\title{Convex hull}
\usage{
convexHull(points)
}
\arguments{
\item{points}{numeric matrix with three columns, one point per row}
}
\value{
A \code{cgalMesh} object. The mesh is triangle.
}
\description{
Mesh of the convex hull of a set of points.
}
\examples{
library(cgalMeshes)
library(rgl)
cube <- rbind(
  c(-1, -1, -1),
  c(-1, -1,  1),
  c(-1,  1, -1),
  c(-1,  1,  1),
  c( 1, -1, -1),
  c( 1, -1,  1),
  c( 1,  1, -1),
  c( 1,  1,  1),
  c( 0,  0,  0)
)
mesh <- convexHull(cube)
# plot
rmesh <- mesh$getMesh()
open3d(windowRect = c(50, 50, 562, 562))
view3d(20, 20)
shade3d(rmesh, color = "orange")
wire3d(rmesh)
}
