% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexHull.R
\name{convexHullsIntersection}
\alias{convexHullsIntersection}
\title{Intersection of convex hulls.}
\usage{
convexHullsIntersection(Points, origin = NULL)
}
\arguments{
\item{Points}{a list of numeric matrices with three columns, each matrix 
represents the points for which a convex hull will be computed}

\item{origin}{either \code{NULL} or a numeric vector of length 3 
corresponding to a point in the intersection; set \code{NULL} if 
you don't know such a point, otherwise give such a point to 
gain speed}
}
\value{
A \code{cgalMesh} object.
}
\description{
Mesh of the intersection of some convex hulls.
}
\examples{
library(cgalMeshes)
library(rgl)
}
