% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crash_centroid.R
\name{crashcentroid}
\alias{crashcentroid}
\title{crash centroid}
\usage{
crashcentroid(x)
}
\arguments{
\item{x}{xxx}
}
\value{
nothing
}
\description{
xxx
}
\examples{
off <- system.file("extdata", "crash_centroid.off", package = "cgalMeshes")
mesh <- cgalMesh$new(off)
print(mesh$centroid())
}
