% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyroTriangle.R
\name{gyroTriangle}
\alias{gyroTriangle}
\title{Hyperbolic triangle}
\usage{
gyroTriangle(A, B, C, s, iterations = 3)
}
\arguments{
\item{A, B, C}{the vertices of the triangle}

\item{s}{hyperbolic curvature, a positive number}

\item{iterations}{number of iterations used to construct the mesh, 
at least 1}
}
\value{
A \code{cgalMesh} object.
}
\description{
Mesh of a hyperbolic triangle.
}
\examples{
library(cgalMeshes)
library(rgl)
mesh <- gyroTriangle(
  c(0, 0, 1), c(1, 0, 0), c(0, 1, 0), s = 0.7
)
mesh$computeNormals()
rmesh <- mesh$getMesh()
open3d(windowRect = 50 + c(0, 0, 512, 512))
view3d(20, 20)
shade3d(rmesh, color = "green")
wire3d(rmesh)
}
