% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isosurface.R
\name{isocontour}
\alias{isocontour}
\title{Title}
\usage{
isocontour(
  voxel,
  x,
  y,
  z,
  isolevel,
  sphereCenter,
  sphereRadius,
  angleBound,
  radiusBound,
  distanceBound
)
}
\arguments{
\item{voxel}{a three dimensional array from which the contour has to 
be calculated}

\item{x, y, z}{locations of the grid points at which values in \code{voxel} 
are measured}

\item{isolevel}{the level at which to construct the isosurface}

\item{sphereCenter, sphereRadius}{center and radius of a bounding sphere}

\item{angleBound}{x}

\item{radiusBound}{x}

\item{distanceBound}{x}
}
\value{
A \code{cgalMesh} object.
}
\description{
Title
}
