% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someMeshes.R
\name{isoCuboidMesh}
\alias{isoCuboidMesh}
\title{Iso-oriented cuboid}
\usage{
isoCuboidMesh(lcorner, ucorner)
}
\arguments{
\item{lcorner}{lower corner, a point whose coordinates must be 
lower than those of the upper corner}

\item{ucorner}{upper corner, a point whose coordinates must be 
greater than those of the lower corner}
}
\value{
A \strong{rgl} mesh, i.e. a \code{mesh3d} object.
}
\description{
Mesh of an iso-oriented cuboid, i.e. a cuboid with 
  edges parallel to the axes.
}
