% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pentagrammicPrism.R
\docType{data}
\name{pentagrammicPrism}
\alias{pentagrammicPrism}
\title{A mesh of a pentagrammic prism}
\format{
A list (\code{vertices}, \code{faces}).
}
\usage{
pentagrammicPrism
}
\description{
A list representing a pentagrammic prism, giving the vertices 
  and the faces; it has 20 vertices, 10 triangular faces, 10 rectangular 
  faces and two pentagonal faces.
}
\keyword{datasets}
