% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonReconstruction.R
\name{PoissonReconstruction}
\alias{PoissonReconstruction}
\title{Poisson surface reconstruction}
\usage{
PoissonReconstruction(
  points,
  normals = "jet",
  neighbors = 12,
  spacing = NULL,
  sm_angle = 20,
  sm_radius = 30,
  sm_distance = 0.375
)
}
\arguments{
\item{points}{numeric matrix which stores the points, one point per row}

\item{normals}{either a numeric matrix which stores the normals, one normal 
per row (i.e. it must have the same size as the \code{points} matrix), or 
one character string, either \code{"jet"} or \code{"pca"}; in this case, 
the normals are computed with the help of the \code{\link{getSomeNormals}} 
function, and you also have to specify the \code{neighbors} argument, 
which is passed to \code{\link{getSomeNormals}}}

\item{neighbors}{argument passed to \code{\link{getSomeNormals}}, see above}

\item{spacing}{size parameter; smaller values increase the precision of the
output mesh at the cost of higher computation time; set to \code{NULL}
(the default) for a reasonable automatic value: an average spacing whose
value will be displayed in a message and that you can also get in the
\code{"spacing"} attribute of the output}

\item{sm_angle}{bound for the minimum facet angle in degrees}

\item{sm_radius}{relative bound for the radius of the surface Delaunay balls}

\item{sm_distance}{relative bound for the center-center distances}
}
\value{
A \code{cgalMesh} object.
}
\description{
Poisson reconstruction of a surface, from a cloud of 3D points.
}
\details{
See \href{https://doc.cgal.org/latest/Poisson_surface_reconstruction_3/index.html}{Poisson Surface Reconstruction}.
}
\examples{
library(cgalMeshes)
library(rgl)
mesh <- PoissonReconstruction(SolidMobiusStrip)
mesh$computeNormals()
rmesh <- mesh$getMesh()
open3d(windowRect = 50 + c(0, 0, 512, 512))
view3d(20, -40, zoom = 0.85)
shade3d(rmesh, color = "darkorange")
wire3d(rmesh)
}
