% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-septuaginta.R
\docType{data}
\name{septuaginta}
\alias{septuaginta}
\title{Mesh of the septuaginta (Da Vinci's 72-sided sphere)}
\format{
A \strong{rgl} mesh (class \code{mesh3d}).
}
\usage{
septuaginta
}
\description{
A \code{mesh3d} object.
 There are some triangle faces and some quad faces.
}
\keyword{datasets}
