% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SolidMobiusStrip.R
\docType{data}
\name{SolidMobiusStrip}
\alias{SolidMobiusStrip}
\title{A point cloud of a solid Möbius strip with noise}
\format{
A matrix with 10000 rows and 3 columns. The 10000 points have been
  sampled at random among more than 400000 points, obtained by the marching
  cubes algorithm, and some noise has been added to these 10000 points.
}
\usage{
SolidMobiusStrip
}
\description{
Some points forming a noisy solid Möbius strip.
}
\references{
\url{http://data.imaginary-exhibition.com/IMAGINARY-Moebiusband-Stephan-Klaus.pdf}
}
\keyword{datasets}
