% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someMeshes.R
\name{sphereMesh}
\alias{sphereMesh}
\title{Sphere mesh}
\usage{
sphereMesh(x = 0, y = 0, z = 0, r = 1, iterations = 3L)
}
\arguments{
\item{x, y, z}{coordinates of the center}

\item{r}{radius}

\item{iterations}{number of iterations (the mesh is obtained by iteratively 
subdividing the faces of an icosahedron)}
}
\value{
A \strong{rgl} mesh (class \code{mesh3d}).
}
\description{
Triangle mesh of a sphere.
}
