% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericalTriangle.R
\name{sphericalTriangle}
\alias{sphericalTriangle}
\title{Spherical triangle}
\usage{
sphericalTriangle(A, B, C, center = c(0, 0, 0), radius = 1, iterations = 4)
}
\arguments{
\item{A, B, C}{the three vertices of the triangle, each given as a pair of 
numbers: the polar angle (between 0 and pi) and the azimuthal angle 
(between 0 and 2pi)}

\item{center}{center of the sphere}

\item{radius}{radius of the sphere}

\item{iterations}{number of iterations used to construct the mesh of 
the sphere}
}
\value{
A \code{cgalMesh} object. The mesh has normals.
}
\description{
Mesh of a spherical triangle.
}
\examples{
library(cgalMeshes)
library(rgl)
# orthant
A <- c(0, pi/2)
B <- c(pi/2, pi/2)
C <- c(0, 0)
mesh <- sphericalTriangle(A, B, C)
rmesh <- mesh$getMesh()
open3d(windowRect = 50 + c(0, 0, 512, 512))
view3d(30, 30)
shade3d(rmesh, color = "red")
wire3d(rmesh)
}
