% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tetrahedraCompound.R
\docType{data}
\name{tetrahedraCompound}
\alias{tetrahedraCompound}
\title{Compound of five tetrahedra}
\format{
A list with two fields: the field \code{Vertices} is a list 
  of five matrices, each one giving the vertices of a tetrahedron; 
  the field \code{rglMeshes} is the list of
  the five \strong{rgl} meshes of the tetrahedra.
}
\usage{
tetrahedraCompound
}
\description{
Five tetrahedra in a pretty configuration. Each tetrahedron
  is centered at the origin.
}
\keyword{datasets}
