% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someMeshes.R
\name{torusMesh}
\alias{torusMesh}
\title{Torus mesh}
\usage{
torusMesh(R, r, p1 = NULL, p2 = NULL, p3 = NULL, nu = 50, nv = 30)
}
\arguments{
\item{R, r}{major and minor radii, positive numbers; \code{R} is 
ignored if \code{p1}, \code{p2} and \code{p3} are given}

\item{p1, p2, p3}{three points or \code{NULL}; if not \code{NULL}, 
the function returns a mesh of the torus whose equator passes 
through these three points and with minor radius \code{r}; if 
\code{NULL}, the torus has equatorial plane z=0 and the 
z-axis as revolution axis}

\item{nu, nv}{numbers of subdivisions, integers (at least 3)}
}
\value{
A triangle \strong{rgl} mesh (class \code{mesh3d}).
}
\description{
Triangle mesh of a torus.
}
\examples{
library(cgalMeshes)
library(rgl)
mesh <- torusMesh(R = 3, r = 1)
open3d(windowRect = 50 + c(0, 0, 512, 512))
view3d(0, 0, zoom = 0.75)
shade3d(mesh, color = "green")
wire3d(mesh)
}
