% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-truncatedIcosahedron.R
\docType{data}
\name{truncatedIcosahedron}
\alias{truncatedIcosahedron}
\title{A mesh of the truncated icosahedron}
\format{
A list with two fields: \code{vertices} and \code{faces}.
}
\usage{
truncatedIcosahedron
}
\description{
A list giving the vertices and the faces of a truncated
  icosahedron. There are some hexagonal faces and some pentagonal faces.
}
\keyword{datasets}
