% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isosurface.R
\name{algebraicMeshesIntersection}
\alias{algebraicMeshesIntersection}
\title{Intersection of algebraic surfaces}
\usage{
algebraicMeshesIntersection(
  polynomials,
  sphereCenter,
  sphereRadius,
  angleBound,
  radiusBound,
  distanceBound,
  errorBound = 0.001
)
}
\arguments{
\item{polynomials}{the polynomials defining the isosurfaces, a list; 
each element of this list can be either a \code{\link[spray]{spray}} 
object or a list with two fields: \code{exponents}, an integer matrix with 
three columns, and \code{coeffs}, a numeric vector; each isosurface is 
then defined by \code{P(x,y,z)=0}, where \code{P} is a polynomial of this 
list}

\item{sphereCenter, sphereRadius}{center and radius of a sphere bounding the 
isosurface; the values of the polynomials at the center of this sphere 
must be less than zero}

\item{angleBound}{lower bound in degrees for the angles of the faces of 
the mesh}

\item{radiusBound}{upper bound for the radii of surface Delaunay balls; 
 a surface Delaunay ball is a ball circumscribing a face and centered 
at the surface}

\item{distanceBound}{upper bound for the distance between the 
circumcenter of a face and the center of the surface Delaunay 
ball of this face}

\item{errorBound}{a relative error bound used in the computations}
}
\value{
A \code{cgalMesh} object.
}
\description{
Computes a mesh of the intersection of some algebraic surfaces 
  (isosurfaces defined by a polynomial).
}
\note{
If either \code{angleBound} is too high, \code{radiusBound} is 
  too small, or \code{distanceBound} is too small, the computation could 
  never finishes. Therefore, it is recommended to start with a small 
  \code{angleBound} and large \code{radiusBound} and \code{distanceBound}, 
  and then to adjust these parameters if the resulting mesh is not nice.
}
\examples{
\donttest{library(cgalMeshes)
library(rgl)
library(spray)
x <- lone(1, 3); y <- lone(2, 3); z <- lone(3, 3)
P1 <-  x^2 + y^2 - z^2 - 1
P2 <-  x^2 - y^2 + z^2 - 1
P3 <- -x^2 + y^2 + z^2 - 1
# compute the mesh of the intersection (trihyperboloid)
mesh <- algebraicMeshesIntersection(
  polynomials = list(P1, P2, P3), 
  sphereCenter = c(0, 0, 0), sphereRadius = 2,
  angleBound    = 10, 
  radiusBound   = 0.008, 
  distanceBound = 0.005,
  errorBound = 5e-6
)
# plot
rmesh <- mesh$getMesh()
open3d(windowRect = c(50, 50, 562, 562))
view3d(20, 20)
bg3d(rgb(54, 57, 64, maxColorValue = 255))
shade3d(rmesh, color = "orangered")}
}
