% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isosurface.R
\name{voxel2mesh}
\alias{voxel2mesh}
\title{Extract mesh from a NifTI file.}
\usage{
voxel2mesh(
  fileName,
  isolevel,
  sphereCenter,
  sphereRadius,
  angleBound,
  radiusBound,
  distanceBound
)
}
\arguments{
\item{fileName}{path to the NifTI file (extension \code{nii} 
or \code{nii.gz})}

\item{isolevel}{the level at which to construct the isosurface}

\item{sphereCenter, sphereRadius}{center and radius of a bounding sphere}

\item{angleBound}{lower bound in degrees for the angles of the faces of 
the mesh}

\item{radiusBound}{upper bound for the radii of surface Delaunay balls; 
 a surface Delaunay ball is a ball circumscribing a face and centered 
at the surface}

\item{distanceBound}{upper bound for the distance between the 
circumcenter of a face and the center of the surface Delaunay 
ball of this face}
}
\value{
A \code{cgalMesh} object.
}
\description{
Make a triangle mesh from a NifTI file and an isovalue.
}
\examples{
library(cgalMeshes)
library(rgl)
teapot <- system.file("extdata", "teapot.nii.gz", package = "cgalMeshes")
mesh <- voxel2mesh(
  teapot, isolevel = 6,
  sphereCenter = c(128, 128, 89), 
  sphereRadius = 200,
  angleBound    = 20, 
  radiusBound   = 5, 
  distanceBound = 5
)
mesh$computeNormals()
rmesh <- rotate3d(mesh$getMesh(), pi, 1, 0, 0)
open3d(windowRect = 50 + c(0, 0, 512, 512))
view3d(0, 0)
shade3d(rmesh, color = "red")
wire3d(rmesh)
}
